reAct-langchain/pyproject.toml
[project]
name = "react-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black>=25.9.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.79",
    "langchain-openai>=0.3.35",
    "python-dotenv>=1.1.1",
]
